Moved action file to function

+++ src/tar-helper.ts
@@ -5,8 +5,11 @@ import * as github from '@actions/github'
 
 // Creates a tar.gzip of the inputs specified in path input or the entire contents
 export async function createTarBall(path: string): Promise<boolean> {
+  const tempDir = '/tmp'
+  const repoNwo: string = process.env.GITHUB_REPOSITORY || '';
+  const repoParse = repoNwo.split('/')
+  const repoName: string = repoParse[1]
   try {
-    const tempDir = '/tmp'
     if (!fs.existsSync(tempDir)) {
       fs.mkdirSync(tempDir)
     }
@@ -17,24 +20,17 @@ export async function createTarBall(path: string): Promise<boolean> {
         'Invalid path. Please ensure the path input has a valid path defined and separated by a space if you want multiple files/folders to be packaged.'
       )
     }
-    const repoNwo: string = process.env.GITHUB_REPOSITORY || '';
-    const repoParse = repoNwo.split('/')
-    const repoName: string = repoParse[1]
 
     if (!fs.existsSync(`${tempDir}/${repoName}`)) {
       fs.mkdirSync(`${tempDir}/${repoName}`);
     }
 
-    // pathArray.forEach(async(filePath) => {
-    //   await exec.exec(`cp -r ${filePath} ${tempDir}/${repoName}`)
-    // })
     for await (const filePath of pathArray) {
       await exec.exec(`cp -r ${filePath} ${tempDir}/${repoName}`)
     }
 
-    const actionFileWithExtension = fs.existsSync('action.yml') ? 'action.yml' : 'action.yaml'
-    if(!isActionYamlPresentInPathSrc(pathArray) && fs.existsSync(actionFileWithExtension) && !fs.existsSync(`${tempDir}/${repoName}/${actionFileWithExtension}`)){
-      await exec.exec(`cp ${actionFileWithExtension} ${tempDir}/${repoName}`)
+    if (!isActionYamlPresentInPathSrc(pathArray)) {
+      copyActionFile(repoName);
     }
 
     const cmd = `tar -czf ${tempDir}/archive.tar.gz -C ${tempDir} ${repoName}`
@@ -42,15 +38,17 @@ export async function createTarBall(path: string): Promise<boolean> {
     await exec.exec(cmd)
     core.info(`Tar ball created.`)
     
-    await exec.exec(`rm -rf ${tempDir}/${repoName}`)
     return true
   } catch (error) {
-  
-    let errorMessage = `Creation of tarball failed! `
-    if (error instanceof Error && error.message)
-      errorMessage += `${error.message}`
-    core.setFailed(errorMessage)
-    return false
+      let errorMessage = `Creation of tarball failed! `
+      if (error instanceof Error && error.message)
+        errorMessage += `${error.message}`
+      core.setFailed(errorMessage)
+      return false
+  } finally { 
+      if(fs.existsSync(`${tempDir}/${repoName}`)){
+        await exec.exec(`rm -rf ${tempDir}/${repoName}`)
+      }
   }
 }
 
@@ -69,15 +67,35 @@ export function isActionYamlPresentInPathSrc(pathArray: string[]): boolean {
 
   // Transform the paths array to remove the traling '/' if it is present in the path input
   pathArray = pathArray.map(e => {
-    if (e.endsWith('/')) return e.slice(0, -1)
+    if (e.endsWith('/')) 
+      return e.slice(0, -1)
     return e
   })
 
   // Returns true as soon as action.y(a)ml is found in any of the paths in the provided path input
   return pathArray.some(filePath => {
     return (
-      fs.existsSync(`${filePath}/action.yml`) ||
-      fs.existsSync(`${filePath}/action.yaml`)
+      fs.existsSync(`${filePath}/action.yml`) || fs.existsSync(`${filePath}/action.yaml`)
     )
   })
+}
+
+export async function copyActionFile(repoName: string): Promise<void> {
+  let actionFileWithExtension = ' ';
+  const tempDir = '/tmp'
+  if(fs.existsSync('action.yml')){
+    actionFileWithExtension = 'action.yml'
+  } else if (fs.existsSync('action.yaml')) {
+      actionFileWithExtension = 'action.yaml'
+  } 
+
+  if(actionFileWithExtension === ' '){
+    throw new Error(
+      `action.y(a)ml not found. Release packages can be created only for action repositories.`
+    )
+  } else if(fs.existsSync(`${tempDir}/${repoName}/${actionFileWithExtension}`)){
+      core.debug(`action.y(a)ml file already exists in ${tempDir}/${repoName}.`)
+  } else {
+      await exec.exec(`cp ${actionFileWithExtension} ${tempDir}/${repoName}`)
+  }
 }
